IdeaCard: move container to separate file
